+++ autoware_lanelet2_extension/lib/deprecated.hpp
@@ -72,6 +72,18 @@ bool getClosestLanelet(
   const lanelet::ConstLanelets & lanelets, const geometry_msgs::msg::Pose & search_pose,
   lanelet::ConstLanelet * closest_lanelet_ptr);
 
+bool isInLanelet(
+  const geometry_msgs::msg::Pose & current_pose, const lanelet::ConstLanelet & lanelet,
+  const double radius);
+
+double getLateralDistanceToCenterline(
+  const lanelet::ConstLanelet & lanelet, const geometry_msgs::msg::Pose & pose);
+
+lanelet::ConstLanelet combineLaneletsShape(const lanelet::ConstLanelets & lanelets);
+
+lanelet::ConstLanelet getExpandedLanelet(
+  const lanelet::ConstLanelet & lanelet_obj, const double left_offset, const double right_offset);
+
 }  // namespace deprecated
 // NOLINTEND(readability-identifier-naming)
