+++ autoware_lanelet2_extension/lib/utilities.cpp
@@ -44,200 +44,6 @@
 #include <utility>
 #include <vector>
 
-namespace impl
-{
-namespace detail
-{
-/// @brief copy the z values between 2 containers based on the 2D arc lengths
-/// @tparam T1 a container of 3D points
-/// @tparam T2 a container of 3D points
-/// @param from points from which the z values will be copied
-/// @param to points to which the z values will be copied
-template <typename T1, typename T2>
-void copyZ(const T1 & from, T2 & to)
-{
-  if (from.empty() || to.empty()) return;
-  to.front().z() = from.front().z();
-  if (from.size() < 2 || to.size() < 2) return;
-  to.back().z() = from.back().z();
-  auto i_from = 1lu;
-  auto s_from = lanelet::geometry::distance2d(from[0], from[1]);
-  auto s_to = 0.0;
-  auto s_from_prev = 0.0;
-  for (auto i_to = 1lu; i_to + 1 < to.size(); ++i_to) {
-    s_to += lanelet::geometry::distance2d(to[i_to - 1], to[i_to]);
-    for (; s_from < s_to && i_from + 1 < from.size(); ++i_from) {
-      s_from_prev = s_from;
-      s_from += lanelet::geometry::distance2d(from[i_from], from[i_from + 1]);
-    }
-    const auto ratio = (s_to - s_from_prev) / (s_from - s_from_prev);
-    to[i_to].z() = from[i_from - 1].z() + ratio * (from[i_from].z() - from[i_from - 1].z());
-  }
-}
-}  // namespace detail
-
-lanelet::ConstLineString3d getClosestSegment(
-  const lanelet::BasicPoint2d & search_pt, const lanelet::ConstLineString3d & linestring)
-{
-  if (linestring.size() < 2) {
-    return lanelet::LineString3d();
-  }
-
-  lanelet::ConstLineString3d closest_segment;
-  double min_distance = std::numeric_limits<double>::max();
-
-  for (size_t i = 1; i < linestring.size(); i++) {
-    lanelet::BasicPoint3d prev_basic_pt = linestring[i - 1].basicPoint();
-    lanelet::BasicPoint3d current_basic_pt = linestring[i].basicPoint();
-
-    lanelet::Point3d prev_pt(
-      lanelet::InvalId, prev_basic_pt.x(), prev_basic_pt.y(), prev_basic_pt.z());
-    lanelet::Point3d current_pt(
-      lanelet::InvalId, current_basic_pt.x(), current_basic_pt.y(), current_basic_pt.z());
-
-    lanelet::LineString3d current_segment(lanelet::InvalId, {prev_pt, current_pt});
-    double distance = lanelet::geometry::distance2d(
-      lanelet::utils::to2D(current_segment).basicLineString(), search_pt);
-    if (distance < min_distance) {
-      closest_segment = current_segment;
-      min_distance = distance;
-    }
-  }
-  return closest_segment;
-}
-
-double getLaneletAngle(
-  const lanelet::ConstLanelet & lanelet, const geometry_msgs::msg::Point & search_point)
-{
-  lanelet::BasicPoint2d llt_search_point(search_point.x, search_point.y);
-  lanelet::ConstLineString3d segment =
-    ::impl::getClosestSegment(llt_search_point, lanelet.centerline());
-  return std::atan2(
-    segment.back().y() - segment.front().y(), segment.back().x() - segment.front().x());
-}
-
-static double getLateralDistanceToCenterline(
-  const lanelet::ConstLanelet & lanelet, const geometry_msgs::msg::Pose & pose)
-{
-  const auto & centerline_2d = lanelet::utils::to2D(lanelet.centerline());
-  const auto lanelet_point = lanelet::utils::conversion::toLaneletPoint(pose.position);
-  return lanelet::geometry::signedDistance(
-    centerline_2d, lanelet::utils::to2D(lanelet_point).basicPoint());
-}
-
-static lanelet::ConstLanelet combineLaneletsShape(const lanelet::ConstLanelets & lanelets)
-{
-  const auto addUniquePoint = [](lanelet::Points3d & points, const lanelet::Point3d & new_point) {
-    constexpr double distance_threshold = 0.01;
-    const auto is_duplicate = std::any_of(
-      points.cbegin(), points.cend(),
-      [&new_point, distance_threshold](const auto & existing_point) {
-        return boost::geometry::distance(existing_point.basicPoint(), new_point.basicPoint()) <=
-               distance_threshold;
-      });
-    if (!is_duplicate) points.emplace_back(new_point);
-  };
-
-  const auto addUniquePoints = [&addUniquePoint](
-                                 lanelet::Points3d & output, const auto & input_points) {
-    std::for_each(
-      input_points.begin(), input_points.end(), [&output, &addUniquePoint](const auto & pt) {
-        addUniquePoint(output, lanelet::Point3d(pt));
-      });
-  };
-
-  lanelet::Points3d lefts, rights, centers;
-  for (const auto & llt : lanelets) {
-    addUniquePoints(lefts, llt.leftBound());
-    addUniquePoints(rights, llt.rightBound());
-    addUniquePoints(centers, llt.centerline());
-  }
-  const auto left_bound = lanelet::LineString3d(lanelet::InvalId, lefts);
-  const auto right_bound = lanelet::LineString3d(lanelet::InvalId, rights);
-  const auto center_line = lanelet::LineString3d(lanelet::InvalId, centers);
-  auto combined_lanelet = lanelet::Lanelet(lanelet::InvalId, left_bound, right_bound);
-  combined_lanelet.setCenterline(center_line);
-  return combined_lanelet;
-}
-
-static lanelet::ConstLanelet getExpandedLanelet(
-  const lanelet::ConstLanelet & lanelet_obj, const double left_offset, const double right_offset)
-{
-  using lanelet::geometry::offsetNoThrow;
-  using lanelet::geometry::internal::checkForInversion;
-
-  const auto & orig_left_bound_2d = lanelet_obj.leftBound2d().basicLineString();
-  const auto & orig_right_bound_2d = lanelet_obj.rightBound2d().basicLineString();
-
-  // Note: The lanelet::geometry::offset throws exception when the undesired inversion is found.
-  // Use offsetNoThrow until the logic is updated to handle the inversion.
-  // TODO(Horibe) update
-  auto expanded_left_bound_2d = offsetNoThrow(orig_left_bound_2d, left_offset);
-  auto expanded_right_bound_2d = offsetNoThrow(orig_right_bound_2d, right_offset);
-
-  rclcpp::Clock clock{RCL_ROS_TIME};
-  try {
-    checkForInversion(orig_left_bound_2d, expanded_left_bound_2d, left_offset);
-    checkForInversion(orig_right_bound_2d, expanded_right_bound_2d, right_offset);
-  } catch (const lanelet::GeometryError & e) {
-    RCLCPP_ERROR_THROTTLE(
-      rclcpp::get_logger("autoware_lanelet2_extension"), clock, 1000,
-      "Fail to expand lanelet. output may be undesired. Lanelet points interval in map data could "
-      "be too narrow.");
-  }
-
-  // Note: modify front and back points so that the successive lanelets will not have any
-  // longitudinal space between them.
-  {  // front
-    const double diff_x = orig_right_bound_2d.front().x() - orig_left_bound_2d.front().x();
-    const double diff_y = orig_right_bound_2d.front().y() - orig_left_bound_2d.front().y();
-    const double theta = std::atan2(diff_y, diff_x);
-    expanded_right_bound_2d.front().x() =
-      orig_right_bound_2d.front().x() - right_offset * std::cos(theta);
-    expanded_right_bound_2d.front().y() =
-      orig_right_bound_2d.front().y() - right_offset * std::sin(theta);
-    expanded_left_bound_2d.front().x() =
-      orig_left_bound_2d.front().x() - left_offset * std::cos(theta);
-    expanded_left_bound_2d.front().y() =
-      orig_left_bound_2d.front().y() - left_offset * std::sin(theta);
-  }
-  {  // back
-    const double diff_x = orig_right_bound_2d.back().x() - orig_left_bound_2d.back().x();
-    const double diff_y = orig_right_bound_2d.back().y() - orig_left_bound_2d.back().y();
-    const double theta = std::atan2(diff_y, diff_x);
-    expanded_right_bound_2d.back().x() =
-      orig_right_bound_2d.back().x() - right_offset * std::cos(theta);
-    expanded_right_bound_2d.back().y() =
-      orig_right_bound_2d.back().y() - right_offset * std::sin(theta);
-    expanded_left_bound_2d.back().x() =
-      orig_left_bound_2d.back().x() - left_offset * std::cos(theta);
-    expanded_left_bound_2d.back().y() =
-      orig_left_bound_2d.back().y() - left_offset * std::sin(theta);
-  }
-
-  const auto toPoints3d = [](const lanelet::BasicLineString2d & ls2d) {
-    lanelet::Points3d output;
-    for (const auto & pt : ls2d) {
-      output.push_back(lanelet::Point3d(lanelet::InvalId, pt.x(), pt.y(), 0.0));
-    }
-    return output;
-  };
-
-  lanelet::Points3d ex_lefts = toPoints3d(expanded_left_bound_2d);
-  lanelet::Points3d ex_rights = toPoints3d(expanded_right_bound_2d);
-  detail::copyZ(lanelet_obj.leftBound3d(), ex_lefts);
-  detail::copyZ(lanelet_obj.rightBound3d(), ex_rights);
-
-  const auto & extended_left_bound_3d = lanelet::LineString3d(lanelet::InvalId, ex_lefts);
-  const auto & expanded_right_bound_3d = lanelet::LineString3d(lanelet::InvalId, ex_rights);
-  const auto & lanelet = lanelet::Lanelet(
-    lanelet_obj.id(), extended_left_bound_3d, expanded_right_bound_3d, lanelet_obj.attributes());
-
-  return lanelet;
-}
-
-}  // namespace impl
-
 namespace lanelet::utils
 {
 namespace
@@ -700,7 +506,7 @@ lanelet::ConstLanelets getExpandedLanelets(
 {
   lanelet::ConstLanelets lanelets;
   for (const auto & llt : lanelet_obj) {
-    lanelets.push_back(impl::getExpandedLanelet(llt, left_offset, right_offset));
+    lanelets.push_back(deprecated::getExpandedLanelet(llt, left_offset, right_offset));
   }
   return lanelets;
 }
@@ -942,7 +748,7 @@ lanelet::ConstLineString3d getClosestSegment(
 lanelet::CompoundPolygon3d getPolygonFromArcLength(
   const lanelet::ConstLanelets & lanelets, const double s1, const double s2)
 {
-  const auto combined_lanelet = ::impl::combineLaneletsShape(lanelets);
+  const auto combined_lanelet = deprecated::combineLaneletsShape(lanelets);
   const auto total_length = lanelet::geometry::length2d(combined_lanelet);
 
   // make sure that s1, and s2 are between [0, lane_length]
@@ -1045,7 +851,7 @@ double getLateralDistanceToClosestLanelet(
 {
   lanelet::ConstLanelet closest_lanelet;
   deprecated::getClosestLanelet(lanelet_sequence, pose, &closest_lanelet);
-  return ::impl::getLateralDistanceToCenterline(closest_lanelet, pose);
+  return deprecated::getLateralDistanceToCenterline(closest_lanelet, pose);
 }
 }  // namespace lanelet::utils
